+++ R/util.R
@@ -60,11 +60,19 @@ to_epochtime <- function(x) {
 # 2) If only `from` is `NULL`, get metrics between 1 day before from `to`, and `to`.
 # 3) If only `to` is `NULL`, get metrics between `from` and now.
 # 4) If both are not `NULL`, get metrics between `from` and `to`.
+#
 # Show a warning if the period is longer than 24 hour since its granularity may be degraded then.
-to_epockperiod <- function(from = NULL, to = NULL) {
+to_epochperiod <- function(from = NULL, to = NULL, .split_request = TRUE) {
   to <- to_epochtime(to %||% Sys.time())
   from <- to_epochtime(from %||% (to - 86400L))
 
-  if(to - from > 86400L) warning("The period is longer than 24 hour; the granularity of metrics may be degraded.")
-  c(from, to)
+  if(.split_request){
+    result <- seq(from, to, by = 86400L)
+    result <- c(result, to)
+  } else {
+    if(to - from > 86400L) warning("The period is longer than 24 hour; the granularity of metrics may be degraded.")
+    result <- c(from, to)
+  }
+
+  result
 }