+++ NEWS.md
@@ -1,6 +1,8 @@
 # datadogr 0.1.0.9000
 
 * Improve the calculation of the time period for metrics.
+* Automatically split the request when the period is longer than a day.
+* `k9_events()` is renamed to `k9_get_events()`.
 
 # datadogr 0.1.0